Day 3 Notes-App Push

+++ day3/django-app.sh
@@ -7,36 +7,54 @@ task
 
 git_clone() {
 	echo "Cloning Django App....."
+if [[ -d django-notes-app ]]; then
+	echo "The repository already exists!!!"
+	echo "Pulling the latest repo...."
+	
+else
+	echo "Cloning Django App....."
 	git clone https://github.com/LondheShubham153/django-notes-app.git
+fi
+echo "*************---------------**************"
 
 }
 
 
 requirements(){
-	echo "Installing requirements"
+echo "*********** Installing Requirements ************"
+if [[ -e /bin/git || -e /bin/nginx ]]; then
+	echo "The services already exists !!!!!!!!"
+else
 	sudo yum install -y docker nginx
+fi
+
 }
 
 required_restarts() {
+echo "*************---------------**************"
+echo "Restarting the services............"
+
  sudo systemctl enable docker
  sudo systemctl enable nginx
+ sudo systemctl restart docker
+ sudo systemctl restart nginx
+
+echo "*************---------------**************" 
 }
 
 deploy(){
+	cd django-notes-app
 	docker build -t notes-app .
 	docker run -d -p 8000:8000 notes-app:latest
 }
 
 echo "********** Deploying The App **********"
-if ! git_clone; then
-	echo "The Repository already exists"
-	cd django-notes-app
-fi	
 
+git_clone
 requirements
 required_restarts
 deploy
-echo "********** Deploying Finsihed  **********"
+echo "********** Deploying Finished  **********"
